gitdb: store Trees as objects in the Blobs table.
The Trees table was kind of pointless; we're never going to do any joins on
it, since we always need to do merge operations on the entire tree object
anyway.  And looking at the plaintext in the Blobs table using sqlite is
actually *easier* than doing it the proper relational way.