Add a new lib/ycoder module that chooses which yaml module to use.
The Debian one is much faster (well, *faster*, though still not fast; good
lord, the pyyaml module sucks a lot!) because it has a C-accelerated
version.  Use it if possible.

Move the pyyaml module we imported earlier into lib/yaml_slow to get it out
of the way.  We still might need this on win32.