cmd/mapi-ls: print full paths, and recurse when no parameters given.
We formerly displayed the folders as a human-readable indented hierarchy;
this was cute, but prevented cut-and-paste.  Really, all anyone wants to do
is use mapi-ls to get a list of folders, then paste that folder name into a
later 'mapi' or 'mapi-items' command line.

Also fix the behaviour that was preventing recursing into the toplevel
"store" objects: they don't have a PR_SUBFOLDERS property, so none of the
children were qualifying.