cmd/fake-merge: a command for marking a branch as merged without merging.
This is good if you want to do an initial import from MAPI into ffcrm, but
not write all the initial entries back the other way.  Mostly just to be
safe.

It's also useful in case you lose your crampi.sqlite3 database: you can mark
things as merged if you're pretty sure they're already up-to-date.  Of
course, that's a little risky, but oh well.