Reload the entries after a merge and before saving.
If we don't do this, any differences between the entrylist stored in the
backend and the one stored in the gitdb will be undone on the *next* import,
which will show a phantom diff, and that phantom diff might get merged into
some other database.  Nasty.

It's entirely possible (and must be acceptable) for a backend to be unable
to support particular attributes of an entry.  As long as we never merge
the *deletion* of those attributes back into some other branch, it'll work
out okay.  But the way it was working, that could happen sometimes since the
attributes would be appearing and disappearing randomly on our branch.

Specifically, this was happening with ffcrm and the addresses[] array up
until I fixed it.